+++ apps/frontend/src/contexts/ThemeContext.ts
@@ -1,0 +1,12 @@
+import { createContext } from 'react';
+
+export type ThemeMode = 'light' | 'dark';
+
+export interface ThemeContextType {
+  mode: ThemeMode;
+  toggleTheme: () => void;
+}
+
+export const ThemeContext = createContext<ThemeContextType | undefined>(
+  undefined
+);
