+++ apps/frontend/src/services/auth.service.ts
@@ -20,7 +20,6 @@ class AuthService {
     if (response.data && response.data.token) {
       localStorage.setItem('token', response.data.token);
     }
-    console.log(response);
     return response.data;
   }
